wired up remove equipment action and reducer
